数値をとりあえずdoubleでやりとりするように戻しておいた。

+++ src/c/frame/frame.cpp
@@ -647,7 +647,7 @@ bool setupJsFrameObject_common(
         return false;
     }
     Nan::Set(jsFrame, Nan::New("type").ToLocalChecked(), Nan::New(type).ToLocalChecked());
-    Nan::Set(jsFrame, Nan::New("pts").ToLocalChecked(), Nan::New((uint32_t)(frame->pts)));
+    Nan::Set(jsFrame, Nan::New("pts").ToLocalChecked(), Nan::New((double)(frame->pts)));
     Nan::Set(jsFrame, Nan::New("timebase").ToLocalChecked(), Nan::New(frame->timebase));
     if(frame->data != NULL) {
         // ここがNULLの場合はぬるぽがでて、動作しないのか・・・
